requirements.txt
Django>=6.0
djangorestframework>=3.14
django-cors-headers>=4.3
drf-spectacular>=0.27
Pillow>=10.0
pillow-avif-plugin>=1.4
python-dotenv>=1.0
django-ratelimit>=4.1
gunicorn>=21.0
